Split column_leaf into files
