+++ src/trace/layers/mod.rs
@@ -4,8 +4,8 @@
 //! elements in the next layer. Similarly, ranges of elements in the layer
 //! itself may correspond to single elements in the layer above.
 
+pub mod column_leaf;
 pub mod ordered;
-pub mod ordered_column_leaf;
 pub mod ordered_leaf;
 // pub mod hashed;
 // pub mod weighted;
